FEAT(client): Add function to pin channels when filtering
Previously, channel could only be explicitly hidden from
the tree view when the channel filter was enabled.

This commit adds the additional option to pin channels
which will be shown in filtered mode, even if there are
no clients connected to them. This is accompanied by
a new method of saving the details in the client sqlite
database: The table 'filtered_channel' was altered to
contain an enum value. Whenever the current user is
in a channel, the channel and all parent channels are
always visible. Otherwise, the channel can be either
hidden, pinned, or behave according to the current
global UI setting when filtering.

Closes #1830

Co-Authored-By: Stefan Hacker <dd0t@users.sourceforge.net>