CHANGE(server): Clear priority speaker flag when switching channels
It is assumed that in most cases the priority speaker flag behavior
should be changed such that it is automatically cleared by the server
once a client switches channels.

This will also prevent malicious clients from giving themselves the
priority speaker flag in a temporary channel and then using that flag
in other channels to annoy users.

Fixes #5334
Partly touches #3584