FIX(server): Fix temp channel permission check logic
Previously in 52e27ef37f a check was implemented to prevent users
in temporary channels from applying mute/deaf/priority speaker.
The reason for this being that temporary channels could, in theory,
be used to circumvent admin mutes.
This check was implemented by testing the channel of the _source user_
for the temporary flag.
This approach is flawed, because it prevents all users (including
admins) in any temporary channel from using these options. (Issue #1877)

This commit introduces a new approach which checks the first
non-temporary parent channel of the target for the MuteDeaf ACL.
It also allows the priority speaker to be set in temporary channels,
because a previous commit added the automatic priority speaker
clearing on switching channels.

The new approach should behave the same as before for mute/deaf
except that priviledged users can now mute/deaf others from
temporary channels (if they have the permission in the channel
above the temporary channel).

Fixes #1877
Partly touches #3584