FEAT(client, server): Add more control over UDP stream.
As requested in #5517 this commit implements the ability to control
the inbound and outbound UDP audio stream by providing additional
options in the network settings. The client and the server works
together to provide these individual controls.

An additional section called "UDP stream mode" has been added.
Each of the options can be selected to specify the behaviour of the UDP
as well as TCP stream.

Also a protobuf message component called "disable_udp_mode" is
added under UserState message which is used to force the server to
use TCP mode only.

Implements #5517

Co-Authored-By: Krzmbrzl <krzmbrzl@gmail.com>