FIX(client): Fix rare double delete bug
AudioOutput keeps track of users and their speech, as well as samples
in qmOutputs. The mix() method will iterate over them and append all
completed outputs to a list 'qlDel'. This list is then - outside
of any synchronization - used to remove entries from qmOutputs.
Previously, other methods were also deleting from qmOutputs directly from
other threads. In rare conditions, when the output just finished and is
added to qlDel and then additionally manually deleted, the pointer could be
subject to a double delete.

This commit tries to mitigate the problem and overhauls the output
deletion by introducing a singal and a slot, delegating the deletion
to the audio thread event queue. A new AudioOutputToken class is
introduced, to prevent direct access to audio buffer pointers.