SQLite: Don't rollback automatically
The function for raising an error inside a custom trigger used to
automatically roll back the currently active transaction. However, if
the transaction is explicitly handled via a RAII wrapper that
automatically rolls back on error, this would lead to a second (invalid)
rollback command to be issued, causing an error.

Instead, we now only abort the current sql statement for which the
trigger failed, but we don't touch the active transaction.