REFAC(client): Rename AudioOutputUser to AudioOutputBuffer
The name AudioOutputUser is a legacy one, or, in case it is meant as
"user of the audio output API", at least slightly misleading.
Since both AudioOutputSpeech and AudioOutputSample inherit AudioOutputUser
it makes more sense to name it as what it actually is: A buffer wrapper.

(cherry picked from commit d79850c94dc1812360633b11ffe393940cac2fcf)