FIX(a11y): Rework autocomplete username. Use TAB to navigate out of chat bar.
The autocomplete function for usernames introduced in 56f5f978c
overwrites the forward-TAB accessibility behaviour. This results in
screen reader users being stuck inside the chat bar having to shift-TAB
or "back TAB" out to reach other user interface modules.

This commit changes the TAB autocomplete behaviour requiring an explicit
"@" in the chatbar for it to trigger. Otherwise it behaves as normal TAB.

See #5811