FIX(a11y): Make users and channels in tree accessible
Previously, screen readers would just read out the string of any selected
item in the tree.

With this new commit, the accessibleText and accessibleDescription are
set such that it is clear whether an item is a channel or a user and
which properties are set for both.

Fixes #2292

(cherry picked from commit 7a9452de67a8aee5e732f854ca413adb233e2df9)