FIX(a11y): Automatically send a focus event to the AudioWizard and CertWizard on page change
Previously, none of the description text of the wizard pages were read
by screen readers. (especially Orca)

To fix this, this commit will focus the dialog itself when
the page changes. This makes the screen reader read the current
page information. It also implements a new function that
finds the next focusable child and sets that in the focus
chain of the wizard as well.