FEAT(client, ui): Allow for max bitrate in client
Via #4700, the server binaries were adapted to allow for an Opus bitrate
of up to 510000 bits per second. The client on the other hand was left
at a max of 192 kb/s.

While this should be more than sufficient for most users, we now adapt
the limit on the client side to (in theory) be able to encode packages
with that 510 kb/s bitrate limit. Outside of very specialized
applications, this should be avoided though.

In order to keep the UI manageable, the respective slider now uses a
logarithmic scale. This also has the benefit that optically, the default
value of 40 kb/s looks pretty good compared to the max. possible value
and therefore novice users hopefully keep that value at its default.

Fixes #5989