FIX(client): Prevent loading unwanted Opus version
This commit ensures that we first try all names for the different base
paths in turn, before we fall back for a generic search for that lib
based only on the given name.
Otherwise, e.g. the system Opus library could be used instead of the
bundled version compiled together with Mumble.