BUILD(cmake): Cross-compile overlay on M1 macs
The overlay doesn't work on ARM Macs, so we disable it on these
machines. However, the CMAKE_SYSTEM_PROCESSOR variable doesn't always
hold the target architecture and as it turned out it contained the host
arch on our M1 macOS builder, causing the overlay to be disabled for our
x86 macOS binaries that were cross-compiled from our M1 machine.

Therefore, we now leverage the system we also use to detect the target
architecture for passing it as a macro to our code.