MAINT: Rewrite changelog script
The new script is more powerful and in particular can compare across
different branches.

Furthermore, it no longer only considers merge commits. Instead, all
commits are screened individually, which should make the script be much
more suitable for situations in which a PR introduces multiple changes
that are not well described by a single commit message (the merge
commit's one).