BUILD(cmake): Fix server-only builds
In #6002 the dbus option was removed and on Linux systems DBus was
unconditionally enabled. However, in the server's cmake file a
copy&paste error was introduced that would produce an error if the
client was not built alongside the server.

This commit fixes this error and therefore makes server-only
builds possible again.