CI(github-actions): Ensure build number tagging
If the actual build is skipped due to a duplicate workflow run, the
build  number script won't run. This can be problematic if this would
have been the only place where the script runs for the given commit as
that'll mean that for this commit no build number will be tagged on our
central server.

Therefore, we now fetch the build number in a separate step that will
not be skipped, if the build itself is skipped.

(cherry picked from commit 472a5f28498f5f7b58b1dedd8c976555682065d7)