Disable migration support for scheme versions < 6
Scheme version 6 was introduced in 2014 and therefore we expect that any
server that is still in use has been updated to a more recent version
between then and now. If not, servers will have to be updated to any
version >= 1.3.x which will perform the necessary migration and
afterwards the DB scheme can be migrated by the current code.

Maintaining backwards compatibility for more than that would only
unnecessarily increase the maintenance burden (for testing the
migrations).