FIX(server): Potential crash due to missing locks
Not taking the qmCache mutex while accessing acCache can lead to
concurrent modification errors, that can cause the server to crash.
Therefore, we now make sure that we hold the mutex before accessing
acCache.

We re-take the lock in every iteration in order to ensure the voice
thread is not stalled by bigger loops over channel or users (the
voice-thread needs to take permissions into account for e.g. linked
channels). For this, we accept the overhead of re-taking the lock in
every iteration in the control thread.