FIX(client, server): Remove unreachable -V flag
Via #5599 support for the --version CLI flag was added. Additionally, a
short-form -V was also added. However, as it turns out Mumble doesn't
consider the casing of CLI arguments and therefore passing -V instead
triggers the code path for -v (small "v") which enables verbose output.

As this option can't be used anyway, it has now been removed and the
help text was adapted accordingly.

Fixes #6037