BUILD(cmake): Ensure ownership of installed INI file
By using systemd-tmpfiles, we can let the system set the ownership of
the installed Mumble server INI file to be root and the group ownership
to be _mumble-server, which is the special user that will start the
server through the systemd service file. The group ownership ensures
that the server has read access to the INI file.