MAINT: Remove system DBus config file
If the config file was installed, then every user on the system running
this DBus service would have unlimited access to the Mumble server. This
is certainly not what we want and given that the DBus API to the server
is considered deprecated anyway, this commit simply removes the config
file from the Mumble source tree.

Fixes #6098