FIX(a11y): Improve accessibility of the edit shortcut dialog
This commit switches the label in the shortcut dialog from
"Add" to "Listening to input" when the button is pressed.
This is supposed to indicate to screen readers that
the dialog is now in a different state, which is actually
listening for new key presses.

(cherry picked from commit ca29279df9cd160561ce1ef60be70ba102008b98)