CHANGE(client): Set min volume for pos. audio to 0
While in 0e4368604974074d5253ad23539f075ececd1e93 it was argued that
this setting should not be set to zero in order to avoid "bug" reports
of audio not working, because the volume of the audio has been zero due
to positional audio, setting this to anything but zero seems like
non-sense for such a feature.
Anything > 0 will lead to the player always being able to hear EVERY
player on the map, regardless of their position, whereas basically the
whole point of positional audio (in a "pure" sense) is to not have this
overlap of a million voices on a full game servers.