FEAT(client): Add popup when mute cue is activated for the first time
As requested in #6094 this commit implements a popup to inform the user
that the mute cue just played while also giving them the option the keep
it enabled or not.

For that new slot functions were added in the MainWindow along with a
signal function in the AudioInput class. These use a new boolean
variable in the Settings Class to check if the cue has been played and a
boolean variable in the Global Class to check if the Configuration UI is
open.

The popup works by emitting a signal from the AudioInput class to
the MainWindow class.

Implements #6094