FIX(client): Fix speex bug, introduced through double fix
As our speex fork contained a fix which was more of a workaround than actually fixing the underlying problem we move back to the upstream xiph/speexdsp repository, as we do not need our modified fork anymore.
The actual problem was fixed through https://github.com/mumble-voip/mumble/pull/6154/files but the combination of both fixes manifested a problem, where the cache was cleared at a point were it should not be cleared yet.