FIX(server,settings,log): announce if settings file is not present
Apparently, QSettings does not care if the specified settings file exists or not.
Therefore, no keys will be loaded into the settings as if the file was empty.
We specifically check for the missing file to not reference a ghost file in the logs.

Fixes #6077

(cherry picked from commit 89ba6acff7feff64521391aa739f5d375e7ed348)