FIX(client,positional-audio): Fix ManualPlacement plugin orientation indicator
Previously, the ManualPlacement plugin for the users position/orientation
was pointing in the wrong (opposite) direction.

This fix swaps the orientation of the sprite.
It also set the default azimuth to 0 instead of 180 to make the
orientation of the listener intuitive for the user by default.
The new default azimuth will not change the logic of the plugin
whatsoever.