FIX(client,positional-audio): Fix positional minimum and maximum distance constraints
8d7e1b52c29 added constraints to the positional distance sliders and spinner
boxes. It was decided that the maximum distance must always be at least
1m larger than the minimum and vice versa.

However, the methods which update the slider value were - incorrectly - using
+1 and -1, which only equates to 0.1m in slide space.
The subsequent update happening in the methods, which update the spinner boxes,
were now using the incorrect 0.1m change in the checks and in turn adding or subtracting 1
in spinner box space (which correctly equates to 1m in spinner box space).

Because of the order of loading operations, only the minimum slider
is affected by the problem: It would reduce by 0.9m everytime the
options dialog was loaded (and saved).

This commit fixes the problem by adding and subtracting 10 in two
of the methods which is 1m in slider space. It also slightly changes
the wording of the comments