FIX(client): Fix using the keyboad to change local volume adjustment
Previously, when using keyboard arrows or the mouse wheel,
the new slider widget action applied the new local volume
but did not save it. That was because it used the sliderReleased
event to save the value, which is not triggered by keyboard or mouse
wheel updates.

We could simply save the local volume on every sliderChange instead of
on sliderRelease, but that would cause many writes to the disk
in a short period of time and possibly lag the client.
Additionally, the ListenerVolumeSlider would spam packets to the
server, as the listener volume is saved server-side.

This commit introduces event filters in addition to the sliderChanged event,
which are observing key releases and mouse wheel events.
It also introduces a delay timer for packets send to the server from the
ListenerVolumeSlider.

Fixes #6211