FIX(client, Markdown): Code blocks end on three backticks (```) without interruption by any other pattern of backticks in the code.
The regex identifying the code blocks of a string in the function `processMarkdownCodeBlock` was here changed from trying to close a code block at its first backtick to only close it by three consecutive backticks.

This allows the use of backticks in code blocks without accidentally cancelling the code block. One example of the content it allows is formatted strings in JavaScript, another is ASCII art containing backticks that do not match the end signature of a code block.

For explanation on how the changed regex operates, see [Regex Match all characters between two strings](https://stackoverflow.com/a/72543781).