FIX(client, ui): Fix log incorrectly scrolling up
When the chat log is detected to be scrolled all the way to the bottom,
it is set to the bottom again when new entries are added to the chat
log.

The detection works properly, but sometimes (especially when an image is
pasted into the chat box) the scroll bar ends up in an incorrect
position.

When the chat log ends up in this incorrect state, subsequent calls to
`scrollLogToBottom` still work correctly, i.e. resizing the window
(which triggers this function) will correctly set the chat log scroll bar
to the bottom.

Therefore, the issue seems to be caused by the scrollbar sometimes being
in an inconsistent state *immediately* after a message is sent, but not
any longer.

My workaround is to add a call to `QApplication::processEvents` before
setting the scrollbar to its maximum position to avoid the inconsistent
state.

As discussed before, this seems to be an upstream bug in Qt, but my
workaround nevertheless seems to fix the issue in my testing.

I have not been able to reproduce the bug locally since applying my
changes.

Fixes #2504