CHANGE(client, ui): remove resizeEvent from LogTextBrowser
The previous implementation of `resizeEvent` included only a call to
`scrollLogToBottom` before calling into the default `resizeEvent`
implementation.

The re-implementation of the method is removed for the following
reasons:

First, the user should be able to resize their chat log without losing
their scrollbar position.

Second, it allows a simpler fix to #2504. The fix entails adding a call
to `QApplication::processEvents` in scrollLogToBottom, but calling this
function from within an event handler causes an endless recursion and
therefore a stack overflow.

This change eliminates an edge case that would have to be accounted for
in the fix.