FIX(a11y): Move all accessible name definitions into XML where applicable
In c73f967711 accessibility names and descriptions were added to Mumble
but they were placed in code.

This commit moves all those texts into the respective .ui file where
applicable. If the accessible name is changed or calculated at runtime,
it obviously needs to still be defined in the source code.