FIX(a11y): Add semantic description to sliders
Some screen readers (notably Orca) are either not capable
of reading semantic values for slides, or Qt prevents them
from doing so.
Sadly, the raw value of a slider e.g. in case of the voice activity
threshsold 0-32767 is pretty much entirely useless for a screen
reader user.

This commit introduces the concept of a semantic slider, which
has a separate accessible text for providing more sensible values
as well as units while still being able to pass the raw slider value
to underlying code.

For screen readers such as Orca, it is a bit hacky as it prepends
the semantic value of the slider to the accessibleName. (Orca does
not read the provided value object) But other than
that, this should work very well for almost all circumstances.

(cherry picked from commit ff4a67612b5654cbf4f2447b693b7e96f2361c3d)