FIX(client): Fix PulseAudio regression by nulling buffer
In b5a67c05f a buffer in the PulseAudio backend was made static, but
at the same time the memset(0) was removed.
In combination, that lead to a bug where the last few frames
of any given audio played back was repeated until something
else filled the buffer.

This commit adds a simple fill to the buffer where the memset(0)
previously was fixing the bug.