FEAT(client): Add "Move To Own Channel" action to user context menu
Previously, there was no way to move users from channel to channel
using only the keyboard. For accessibility reasons, that was not ideal.

This commit adds a new action to the user context menu allowing
to move the selected user to your own channel (given enough
permissions).

While this is a decent workaround, the end goal will be to remove
this action and replace it with a dedicated dialog to choose
the channel a user will be moved to.

See #4642

(cherry picked from commit f63b7b42db3a16424aa85abbe5af35c0166ea868)