FEAT(client): disable pipewire input stream while muted
When the user is muted and/or suppressed in Mumble, there is no reason to keep the audio stream running (except for the mute cue).

Disabling/suspending the pipewire stream has the advantage of allowing upstream nodes to suspend themselves as well. When using resource-heavy audio filters like e.g. DeepFilterNet, this can save significant amounts of CPU time.