FIX(client): Fix incorrect WASAPI sound device handling
In b5a67c0 a refactor took place which removed some compiler warnings.
In WASAPI.cpp a STACKVAR was swapped for a static std::vector,
which led to some unforeseen consequences and total audio input
failure on some systems.

This commit removes the static modifier as it is not needed there
anyway.

Fixes #6356