DOCS(docker): Fix file and doc references
PR #5838 commit 81bb8f70 removed the `docker-compose.yml` file in this repository.
It points to our mumble-docker repository for docker related things.
In that repository, we do not currently have a default or example docker compose configuration file.
The README in that repository does have a description for it, so we link to that now.

Update the relative file path reference `murmur.ini` -> `mumble-server-ini`.

Update `database` path 'murmur' -> 'mumble-server'.
I did not test or verify this change. Even if it is incorrect or does not work, I assume it to be *more correct*. Even if nobody chooses to test or verify it in this doc-update-changeset, I think that's preferable.