DOCS(docker): Drop outdated docker compose documentation
Docker documentation should be on our website (administrator/hoster documentation) or our [mumble-docker] repository.
This repository /docs/ should only contain developer documentation and scoped to this source/main project.

---

PR #5838 commit 81bb8f7 removed the docker-compose.yml file in this repository.

The `murmur.ini` reference is outdated. It was moved and renamed to `umble-server-ini`

The `database` path is outdated ('murmur' name).

---

Replaces/Closes PR #6387 which intended to fix the documentation.

[mumble-docker]: https://github.com/mumble-voip/mumble-docker/